# pagecache benchmarks